TransactionUI can now change the button texts and functions based on the mode selected.
btnSearch(The button labelled search in the default mode) can now perform all its functions.

Rewrote the display message function to now require the title of the dialog box to be passed to the function. This fixed the Issue I was dealing with were it crashed whenever I passed the Operation Flags into display_message.

Will remove the constants later and fix all the calls to display_message by the LoginUI.

Added a new dialog box function to the messages module that returns a flag/value based on what the user clicked. for now it can only do yes and no action confirmation messages but this will be expanded upon later.

display_operation_status function in the messages module has been removed because it is not needed by any more after the display_message refactor

Removed the Flags and pending tasks comments in messages module pending to the bugs of display_message and the planned functions of display_operation_status functions